UIIN-2062 Rename InstanceFilterBrowse
