--- src/components/InstanceFilters/InstanceBrowseFilters/index.js
@@ -1,1 +1,0 @@
-export { default } from './InstanceFilterBrowse';
